# openapi/006-identity-and-authorization.yaml
openapi: 3.0.3
info:
  title: Tovu Admin API — Identity and Authorization
  version: "1.0.0"
  description: Hand-verified against the current route handlers, serializers, and identity-service validation code on 2026-08-24.
servers: [{ url: / }]
...
security: [{ adminSessionCookie: [] }, { apiKeyAuthorization: [] }]
...
paths:
...
  /api/admin/v1/api-keys:
    post:
      operationId: issue_api_key
      tags: [api-keys]
      summary: Issue an API key bound to a principal
      description: >-
        Issues a key bound to a grantless kind='api_key' principal. The requested policyIds are
        SNAPSHOTTED at issuance — their permission rows are copied field-identically (resourceType
        and constraintJson preserved per row, never widened) into a new immutable isFrozen,
        machine-owned policy attached to that principal — so a later write to a source policy cannot
        grow an already-issued key. Every permission the snapshot would confer must be held
        UNCONSTRAINED by the caller (INV-07) or issuance is refused with no rows written at all,
        including no orphan snapshot policy. Requires apikey.manage. Handler:
        src/server/routes/admin/api-keys/issue.ts.
      security: [{ adminSessionCookie: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/IssueApiKeyRequest' } } }
      responses:
        '201': { description: 'Raw key shown once; only a salted scrypt digest of its secret half is persisted, so no later read of any table can reproduce it.', content: { application/json: { schema: { $ref: '#/components/schemas/ApiKeyIssueResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/ForbiddenApiKeyIssuance' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/admin/v1/workspaces/{workspaceId}/policies/{policyId}/permissions:
    post:
      operationId: write_policy_permission
      tags: [policies]
      summary: Add a permission to a mutable policy
      parameters: [{ $ref: '#/components/parameters/WorkspaceId' }, { $ref: '#/components/parameters/PolicyId' }]
      # Drift: api.spec.md permits null resourceType/constraintJson; the handler coerces each null to the literal string "null".
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/WritePolicyPermissionRequest' } } } }
      responses:
        '201': { description: Created policy permission, content: { application/json: { schema: { $ref: '#/components/schemas/PolicyPermissionResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationOrUnknownPermission' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/ForbiddenOrGrantClamp' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    adminSessionCookie: { type: apiKey, in: cookie, name: tovu_session }
    # Declared as an explicit header credential rather than `type: http, scheme: bearer` because the
    # runtime accepts TWO spellings of the same header and an http/bearer scheme could express only
    # one of them. Keeping it honest about both beats picking the tidier declaration.
    apiKeyAuthorization: { type: apiKey, in: header, name: Authorization, description: 'An issued API key, sent as `Bearer <raw-key>` or `ApiKey <raw-key>` (case-insensitive scheme). Raw keys are emitted once, by issue_api_key, and are never returned by any other endpoint.' }
  parameters:
    WorkspaceId: { name: workspaceId, in: path, required: true, description: Compared directly with the server-bound workspace; no UUID-format validation is performed., schema: { type: string } }
...
    PolicyId: { name: policyId, in: path, required: true, description: Passed to the identity service without route-level UUID validation., schema: { type: string } }
  schemas:
...
    WritePolicyPermissionRequest: { type: object, required: [permission], properties: { permission: { type: string }, resourceType: { type: string, description: null is coerced to the string 'null'. }, constraintJson: { type: string, description: null is coerced to the string 'null'. } } }
...
    PolicyPermission: { type: object, required: [id, workspaceId, policyId, permission, resourceType, constraintJson], properties: { id: { type: string }, workspaceId: { type: string }, policyId: { type: string }, permission: { type: string }, resourceType: { type: string, nullable: true }, constraintJson: { type: string, nullable: true } } }
...
    PolicyPermissionResponse: { type: object, required: [policyPermission], properties: { policyPermission: { $ref: '#/components/schemas/PolicyPermission' } } }
...
    IssueApiKeyRequest: { type: object, required: [principalId, label, policyIds], properties: { principalId: { type: string, description: "Must reference a kind='api_key' principal in the caller's workspace that is still grantless. A user/system target is rejected 400 (a key must not inherit a privileged principal's grants); an already-granted target is rejected 400; an unknown one is 404." }, label: { type: string, minLength: 1, maxLength: 255, description: Bounds are measured after trimming. }, policyIds: { type: array, minItems: 1, items: { type: string }, description: "Source policies to snapshot. A policy carrying the owner wildcard '*' is rejected 400 — an api-key snapshot may never carry it. An unknown policy is 404." }, expiresAt: { type: string, format: date-time, nullable: true, description: "Optional; null, absent, or empty string all mean no expiry. Canonicalized to UTC on write and compared as an instant, so an offset-bearing value is honored exactly." } } }
    ApiKeyIssueResponse: { type: object, required: [apiKey], properties: { apiKey: { type: object, required: [id, rawKey, principalId, label, expiresAt], properties: { id: { type: string }, rawKey: { type: string, description: "The full key, shaped tovu_ak_<12 hex>.<43 base64url chars> — a non-secret prefix used as the lookup handle, plus 256 bits of CSPRNG secret. Shown once, here, and never returned by any other endpoint." }, principalId: { type: string }, label: { type: string }, expiresAt: { type: string, format: date-time, nullable: true } } } } }
    ErrorResponse: { type: object, required: [error], properties: { error: { type: string }, code: { type: string }, details: { type: object, additionalProperties: true } } }
  responses:
    Unauthenticated: { description: Missing, invalid, expired, revoked, or disabled session; login also uses this for invalid credentials., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    ForbiddenOrGrantClamp: { description: Caller lacks role.manage or the requested grant exceeds their unconstrained permissions., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    ForbiddenApiKeyIssuance: { description: "Everything ForbiddenApiKeyManagement covers, plus GRANT_EXCEEDS_ISSUER when the requested snapshot would confer a permission the caller does not hold unconstrained (INV-07); details.offendingPermissions names each one and no rows are written.", content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    NotFound: { description: Missing target. Some assignment/attachment handlers omit code; other handlers send RESOURCE_NOT_FOUND., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    ValidationError: { description: Invalid input or immutable/invalid target., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    ValidationOrUnknownPermission: { description: Validation failure or an unregistered permission (PERMISSION_UNKNOWN)., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    InternalError: { description: "Internal error. Runtime body is {error: 'internal error'} with no code.", content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }